@@ -16,32 +16,32 @@
     "@tauri-apps/api": "^2.0.3",
     "@tauri-apps/plugin-clipboard-manager": "^2.0.0",
     "@tauri-apps/plugin-dialog": "~2.0.1",
-    "@tauri-apps/plugin-fs": "~2.0.1",
+    "@tauri-apps/plugin-fs": "~2.0.2",
     "@tauri-apps/plugin-os": "~2.0.0",
     "@tauri-apps/plugin-shell": "^2.0.1",
     "execution-time-decorators": "^2.0.3"
   },
   "devDependencies": {
     "@sveltejs/adapter-static": "^3.0.6",
-    "@sveltejs/kit": "^2.7.3",
+    "@sveltejs/kit": "^2.7.7",
     "@sveltejs/vite-plugin-svelte": "^3.1.2",
     "@tailwindcss/typography": "^0.5.15",
     "@tauri-apps/cli": "^2.0.4",
     "autoprefixer": "^10.4.20",
-    "bits-ui": "0.21.16",
+    "bits-ui": "1.0.0-next.45",
     "clsx": "^2.1.1",
     "cmdk-sv": "^0.0.18",
     "lucide-svelte": "^0.453.0",
     "mode-watcher": "^0.4.1",
     "paneforge": "^0.0.6",
-    "svelte": "^4.2.19",
+    "svelte": "^5.1.10",
     "svelte-check": "^4.0.5",
     "svelte-headless-table": "^0.18.3",
     "svelte-sonner": "^0.3.28",
     "tailwind-merge": "^2.5.4",
     "tailwind-variants": "^0.2.1",
     "tailwindcss": "^3.4.14",
-    "tslib": "^2.8.0",
+    "tslib": "^2.8.1",
     "typescript": "^5.6.3",
     "vaul-svelte": "^0.3.2",
     "vite": "^5.4.10"
